# Optimize queries